user micro refactoring
refactoring User Micro service layer with only 1 service: UserSolutionServiceImp
