--- itachallenge-user/src/main/java/com/itachallenge/user/service/IServiceChallengeStatistics.java
@@ -1,14 +1,0 @@
-package com.itachallenge.user.service;
-
-import com.itachallenge.user.dtos.ChallengeStatisticsDto;
-import reactor.core.publisher.Mono;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface IServiceChallengeStatistics {
-    Mono<List<ChallengeStatisticsDto>> getChallengeStatistics(List<UUID> challengeIds);
-    Mono<Long> getBookmarkCountByIdChallenge(UUID idChallenge);
-
-    Mono<Float> getChallengeUsersPercentage(UUID idChallenge);
-}
