+++ itachallenge-user/src/main/java/com/itachallenge/user/service/IUserSolutionService.java
@@ -1,10 +1,7 @@
 package com.itachallenge.user.service;
 
 import com.itachallenge.user.document.UserSolutionDocument;
-import com.itachallenge.user.dtos.SolutionUserDto;
-import com.itachallenge.user.dtos.UserScoreDto;
-import com.itachallenge.user.dtos.UserSolutionDto;
-import com.itachallenge.user.dtos.UserSolutionScoreDto;
+import com.itachallenge.user.dtos.*;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -17,5 +14,8 @@ public interface IUserSolutionService {
     Mono<UserSolutionScoreDto> addSolution(UserSolutionDto userSolutionDto);
     Mono<UserSolutionDocument> markAsBookmarked(String uuidChallenge, String uuidLanguage, String uuidUser, boolean bookmarked);
     Flux<UserSolutionDto> showAllUserSolutions(UUID userUuid);
+    Mono<List<ChallengeStatisticsDto>> getChallengeStatistics(List<UUID> challengeIds);
+    Mono<Long> getBookmarkCountByIdChallenge(UUID idChallenge);
+    Mono<Float> getChallengeUsersPercentage(UUID idChallenge);
 
 }
