+++ itachallenge-user/src/test/java/com/itachallenge/user/controller/UserControllerTest.java
@@ -3,15 +3,12 @@ package com.itachallenge.user.controller;
 import com.itachallenge.user.document.UserSolutionDocument;
 import com.itachallenge.user.dtos.*;
 import com.itachallenge.user.exception.UnmodifiableSolutionException;
-import com.itachallenge.user.dtos.*;
-import com.itachallenge.user.service.IServiceChallengeStatistics;
 import com.itachallenge.user.service.IUserSolutionService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -24,12 +21,10 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
-
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
@@ -54,8 +49,6 @@ class UserControllerTest {
 
     @MockBean
     IUserSolutionService userSolutionService;
-    @MockBean
-    IServiceChallengeStatistics statisticsService;
 
     @BeforeEach
     public void setUp() {
@@ -161,7 +154,7 @@ class UserControllerTest {
         String URI_TEST = "/bookmarks/{idChallenge}";
         Long testCount = 1L;
 
-        when(statisticsService.getBookmarkCountByIdChallenge(VALID_MONGO_UUID))
+        when(userSolutionService.getBookmarkCountByIdChallenge(VALID_MONGO_UUID))
                 .thenReturn(Mono.just(testCount));
 
         webTestClient.get()
@@ -192,7 +185,7 @@ class UserControllerTest {
         float percentage = 75.0f;
         ChallengeUserPercentageStatisticDto expectedDto = new ChallengeUserPercentageStatisticDto(challengeId, percentage);
 
-        when(statisticsService.getChallengeUsersPercentage(challengeId)).thenReturn(Mono.just(percentage));
+        when(userSolutionService.getChallengeUsersPercentage(challengeId)).thenReturn(Mono.just(percentage));
 
         webTestClient.get()
                 .uri(CONTROLLER_URL + "/statistics/percent/{idChallenge}", challengeId)
@@ -272,6 +265,7 @@ class UserControllerTest {
                     verify(userSolutionService).addSolution(userSolutionDto);
                 });
     }
+
     @DisplayName("UserDocumentControllerTest - addSolution - return 400 BAD REQUEST and don't save if dto is invalid")
     @Test
     void addSolutionIfInvalidValuesThenBadRequest_test() {
@@ -296,6 +290,7 @@ class UserControllerTest {
             verifyNoInteractions(userSolutionService);
         }
     }
+
     @DisplayName("UserDocumentControllerTest - addSolution - return 409 CONFLICT if Service returns UnmodifiableSolutionException")
     @Test
     void addSolutionServiceThrowsExceptionInternalServerError_test() {
@@ -359,8 +354,4 @@ class UserControllerTest {
                     assertEquals("Sample Solution", solution.getSolutionText());
                 });
     }
-
 }
-
-
-
